Article Plugin Version 2
This article plugin does the following:
- you select the default article and whether a query of ?art=#
overrides the default.

Once you have generated a magazine detail page or a general detail page
for multiple magazines. You enter that URL in the Publication record
for that magazine.

The plugin checks if the URL query can override the default. If it can
and if the publication of the article has a detail view url, it uses
that url. If it doesn’t then it throws a 404.