bug fix for personalized logged in user with primary and secondary discipline
There was a bug where if a visitor had both a primary and secondary
discipline it would try to append the secondary records to the ones for
the primary.  I could have just changed the append to add and it would
work but I thought it was better to change the query to be
discipine__in both instead.  Please see comments in code.